Implemented API access for parcels
This portion of the API wrapper is implemented in much the same way as
the address portion. The only main differences are that parcels do not
get validated by the API, and the endpoints are different.

It certainly stands to reason by looking at the documentation and
noticing most of the resources allow the same methods in much the same
way, there will be a lot of repetition across most of the API wrapper.

Perhaps a default behavior can be used in each of the API portions,
simply overriding where needed.