Improved codegen of Resource
Now allowing authors to specify which methods a resource should supply
defaults for by passing `only` or `except` opts to the use statement.

For example:

```
defmodule Shippo.SomeResource do
  use Shippo.Resource, only: [get: 1]
end
```

Will only generate a default implementation for the get API method.

This allows the library to closer follow the API while keeping
boilerplate to a minimum, and hopefully bugs.